requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-richtextfield==1.6.2
django-unfold==0.80.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
